+++ actionlist/app/reducers/actionItems.js
@@ -1,25 +1,7 @@
 /* eslint arrow-body-style: ["error", "as-needed", { "requireReturnForObjectLiteral": true }] */
 import { handleActions } from 'redux-actions';
 import { TOGGLE_DRAWER, SORT_ACTION_LIST, FILTER_ACTION_LIST, RESET_FILTERS, SELECT_DROPDOWN_OPTION, RESET_PREFERENCES } from '../actions/types';
-import { processInstances, sortTypes, filterStatus, Colors } from '../lib/commons';
-
-export const defaultState = {
-  dataSource: processInstances,
-  drawerExpanded: false,
-  filterStatus,
-  optionSelected: '',
-  dropdownColors: {
-    Saved: Colors.white,
-    Initiated: Colors.white,
-    Disapproved: Colors.white,
-    Enroute: Colors.white,
-    Approved: Colors.white,
-    Final: Colors.white,
-    Processed: Colors.white,
-    Exception: Colors.white,
-    Cancel: Colors.white,
-  },
-};
+import { defaultState, sortTypes, filterStatus } from '../lib/commons';
 
 const sortActionList = (state, action) => {
   const sortedByCriteria = [...state.dataSource];
